awscli cached

+++ scripts/install-cloud-cli.sh
@@ -25,11 +25,28 @@ function installAWSCLI () {
     if [[ -z $isexists ]]
     then
         printf "\naws cli not found. Installing...\n"
-        curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o /tmp/awscliv2.zip
-        cd /tmp/
-        unzip awscliv2.zip
+        if [[ ! -d $HOME/awscli ]]
+        then
+            mkdir awscli
+        fi
+        local removezip=''
+        if [[ ! -d $HOME/awscli/aws ]]
+        then
+            printf "\naws cli binary not found. Downloading...\n"
+            curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o $HOME/aws/awscliv2.zip
+            cd $HOME/awscli/
+            unzip awscliv2.zip
+            $removezip=true
+        fi
+        printf "\ninstalling cli...\n"
+        cd $HOME/awscli/
         ./aws/install
-        rm -rf awscliv2.zip
+        if [[ $removezip == true ]]
+        then
+            rm -rf awscliv2.zip
+        fi
+        cd ~
+        aws --version
     else
         printf "\naws cli found. No need to install new.\n"
         return 0
